# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Curso Api Lumen",
        "description": "Lumen Swagger Api",
        "license": {
            "name": "my api"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Api del curso de Lumen"
        }
    ],
    "paths": {
        "/admin/role/assignPermission/{id}": {
            "put": {
                "tags": [
                    "Role"
                ],
                "summary": "Assing Permission to Role",
                "operationId": "Update Permissions",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "permissions": {
                                        "description": "Permissions Array",
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            }
        },
        "/admin/role/getPermission/{id}": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Show Role Permission",
                "description": "Returns permissions of role",
                "operationId": "Show Permission",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "authorization": {
                "type": "apiKey",
                "name": "authorization",
                "in": "header"
            }
        }
    }
}